+++ packages/sui-bundler/bin/sui-bundler-build.js
@@ -73,9 +73,7 @@ webpack(config).run((error, stats) => {
 
   console.log(
     chalkSuccess(
-      `Your app is compiled in ${
-        process.env.NODE_ENV
-      } mode in /public. It's ready to roll!`
+      `Your app is compiled in ${process.env.NODE_ENV} mode in /public. It's ready to roll!`
     )
   )