+++ packages/sui-ssr/server/ssr/index.js
@@ -108,12 +108,23 @@ export default (req, res, next) => {
         return next(err)
       }
 
+      const {initialProps, reactString, performance} = initialData
+
+      // The __HTTP__ object is created before earlyFlush is applied
+      // to avoid unexpected behaviors
+
+      const {__HTTP__} = initialProps
+      if (__HTTP__) {
+        const {redirectTo} = __HTTP__
+        if (redirectTo) {
+          return res.redirect(HTTP_PERMANENT_REDIRECT, redirectTo)
+        }
+      }
+
       // Flush now if early-flush is disabled
       if (!req.app.locals.earlyFlush) {
         initialFlush(res)
       }
-
-      const {initialProps, reactString, performance} = initialData
 
       // The first html content has the be set after any possible call to next().
       // Otherwise some undesired/duplicated html could be attached to the error pages if an error occurs