+++ packages/sui-dashboard/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+<a name="1.4.0"></a>
+# 1.4.0 (2019-10-02)
+
+
+### Features
+
+* add new projects to the dashboard ([9c36cd3](https://github.com/SUI-Components/sui/commit/9c36cd3))
+
+
+
 <a name="1.3.0"></a>
 # 1.3.0 (2019-02-14)
 
@@ -19,6 +29,23 @@ All notable changes to this project will be documented in this file.
 ### Bug Fixes
 
 * binary name ([e9bfbc2](https://github.com/SUI-Components/sui/commit/e9bfbc2))
+
+
+
+<a name="1.1.0"></a>
+# 1.1.0 (2018-12-12)
+
+
+### Bug Fixes
+
+* fix package name ([38550e1](https://github.com/SUI-Components/sui/commit/38550e1))
+
+
+### Features
+
+* first commit ([ed34f18](https://github.com/SUI-Components/sui/commit/ed34f18))
+* remove excel feature and show better stats ([9138467](https://github.com/SUI-Components/sui/commit/9138467))
+* save report stats to google drive ([89ba52f](https://github.com/SUI-Components/sui/commit/89ba52f))